리팩토링 및 테스터 코드 완성

+++ src/repositories/posts.repository.js
@@ -1,3 +1,5 @@
+import { CustomError } from "../errors/customError.js";
+
 export class PostsRepository {
   constructor(prisma) {
     this.prisma = prisma
@@ -25,21 +27,55 @@ export class PostsRepository {
 
   getOnePost = async (postId) => {
     const post = await this.prisma.posts.findUnique({ where: { postId: +postId } });
+    if (!post) {
+      throw new CustomError(404,'게시글이 존재하지 않습니다')
+    };
     return post;
   };
 
-  updatePost = async (postId, title, content) => {
+  updatePost = async (userId,postId,title,content) => {
+    // step1: 먼저, 에러체크
+    const post = await this.prisma.posts.findUnique({ where: { postId: +postId } });
+    if (!post) {
+      throw new CustomError(404,'게시글이 존재하지 않습니다')
+    };
+    if (post['UserId'] !== userId) { // '!post'조건보다 아래에 있어야 에러x
+      throw new CustomError(403,'게시글 수정 권한이 없습니다')
+    };
+
+    // step2: 비로소, update 수행
     const updatedPost = await this.prisma.posts.update({
       where: { postId: +postId },
       data: { title, content }
     });
-    return updatedPost;
+
+    if (!updatedPost) {
+      throw new CustomError(401,'게시글이 정상적으로 수정되지 않았습니다')
+    };
+    
+    return updatedPost
   };
 
-  deletePost = async (postId) => {
+  deletePost = async (userId,postId) => {
+    // step1: 먼저, 에러체크
+    const post = await this.prisma.posts.findUnique({ where: { postId: +postId } });
+    if (!post) {
+      throw new CustomError(404,'게시글이 존재하지 않습니다')
+    };
+    if (post.UserId !== userId) { // '!post'조건보다 아래에 있어야 에러x
+      throw new CustomError(403,'게시글 삭제 권한이 없습니다')
+    };
+
+    // step2: 비로소, delete 수행
     const deletedPost = await this.prisma.posts.delete({
       where: { postId: +postId }
     });
-    return deletedPost;
+
+    if (!deletedPost) {
+      throw new CustomError(401,'게시글이 정상적으로 삭제되지 않았습니다')
+    };
+
+    return deletedPost // test용 return
   };
+
 };
